examples/libjpeg: correct casting

+++ examples/libjpeg/persistent-jpeg.c
@@ -65,7 +65,7 @@ int LLVMFuzzerTestOneInput(const uint8_t* buf, size_t len) {
     jpeg_read_header(&cinfo, TRUE);
 
     /* Limit total number of pixels to decode to 50M */
-    uint64_t total_pix = (uint64_t)cinfo.output_height * (uint32_t)cinfo.output_width;
+    uint64_t total_pix = (uint64_t)cinfo.output_height * (uint64_t)cinfo.output_width;
     if (total_pix > max_total_pixels) {
         goto out;
     }
